+++ OkayegTeaTime/Files/Jsons/Settings/UserLists.cs
@@ -4,11 +4,11 @@ namespace OkayegTeaTime.Files.Jsons.Settings;
 
 public class UserLists
 {
-    public List<int> Owners { get; set; }
+    public List<long> Owners { get; set; }
 
-    public List<int> Moderators { get; set; }
+    public List<long> Moderators { get; set; }
 
-    public List<int> IgnoredUsers { get; set; }
+    public List<long> IgnoredUsers { get; set; }
 
-    public List<int> SecretUsers { get; set; }
+    public List<long> SecretUsers { get; set; }
 }
